+++ lib_log_utils/__init__conf__.py
@@ -1,20 +1,24 @@
-name = 'lib_log_utils'
-title = 'colored log messages and banners from commandline and python'
-version = 'v1.4.10'
-url = 'https://github.com/bitranox/lib_log_utils'
-author = 'Robert Nowotny'
-author_email = 'bitranox@gmail.com'
-shell_command = 'log_util'
+# CONF
+
+name = "lib_log_utils"
+title = "colored log messages and banners from commandline and python"
+version = "2022-03-25"
+url = "https://github.com/bitranox/lib_log_utils"
+author = "Robert Nowotny"
+author_email = "bitranox@gmail.com"
+shell_command = "log_util"
 
 
 def print_info() -> None:
-    print("""\
+    print(
+        """\
 
 Info for lib_log_utils:
 
     colored log messages and banners from commandline and python
 
-    Version : v1.4.10
+    Version : 2022-03-25
     Url     : https://github.com/bitranox/lib_log_utils
     Author  : Robert Nowotny
-    Email   : bitranox@gmail.com""")
+    Email   : bitranox@gmail.com"""
+    )