+++ lib_log_utils/log_levels.py
@@ -6,23 +6,23 @@ NOTICE: int = 25
 SUCCESS: int = 35
 
 # noinspection PyTypeHints
-logging.SPAM: int = SPAM                                     # type: ignore
+logging.SPAM: int = SPAM  # type: ignore
 # noinspection PyTypeHints
-logging.VERBOSE: int = VERBOSE                               # type: ignore
+logging.VERBOSE: int = VERBOSE  # type: ignore
 # noinspection PyTypeHints
-logging.NOTICE: int = NOTICE                                 # type: ignore
+logging.NOTICE: int = NOTICE  # type: ignore
 # noinspection PyTypeHints
-logging.SUCCESS: int = SUCCESS                               # type: ignore
+logging.SUCCESS: int = SUCCESS  # type: ignore
 
-logging._levelToName[logging.SPAM] = 'SPAM'             # type: ignore
-logging._levelToName[logging.VERBOSE] = 'VERBOSE'       # type: ignore
-logging._levelToName[logging.NOTICE] = 'NOTICE'         # type: ignore
-logging._levelToName[logging.SUCCESS] = 'SUCCESS'       # type: ignore
+logging._levelToName[logging.SPAM] = "SPAM"  # type: ignore
+logging._levelToName[logging.VERBOSE] = "VERBOSE"  # type: ignore
+logging._levelToName[logging.NOTICE] = "NOTICE"  # type: ignore
+logging._levelToName[logging.SUCCESS] = "SUCCESS"  # type: ignore
 
-logging._nameToLevel['SPAM'] = logging.SPAM             # type: ignore
-logging._nameToLevel['VERBOSE'] = logging.VERBOSE       # type: ignore
-logging._nameToLevel['NOTICE'] = logging.NOTICE         # type: ignore
-logging._nameToLevel['SUCCESS'] = logging.SUCCESS       # type: ignore
+logging._nameToLevel["SPAM"] = logging.SPAM  # type: ignore
+logging._nameToLevel["VERBOSE"] = logging.VERBOSE  # type: ignore
+logging._nameToLevel["NOTICE"] = logging.NOTICE  # type: ignore
+logging._nameToLevel["SUCCESS"] = logging.SUCCESS  # type: ignore
 
 
 def get_log_level_from_str(log_level_str: str) -> int:
@@ -46,7 +46,7 @@ def get_log_level_from_str(log_level_str: str) -> int:
         pass
 
     try:
-        log_level = logging._nameToLevel[log_level_str.upper()]
+        log_level = logging._nameToLevel[log_level_str.upper()]  # noqa
         return log_level
     except KeyError:
         raise ValueError(f'can not detect log level from string "{log_level_str}"')